requirements.txt
bcrypt==3.1.0
cffi==1.7.0
coverage==4.2
Django==1.7
django-braces==1.8.1
django-registration-redux==1.4
Pillow==3.3.0
pycparser==2.14
six==1.10.0
